==> cascade/2_cascade_array.hits.ipynb <==
source('cascade_helper.R')


### list of parameters and filenames

wd <- getwd()

p_thr <- 5e-9
maf_min_thr <- 1e-4

annot.tbl <- '@@@@@@/private_data/ukbb/variant_filtering/variant_filter_table.tsv.gz'

cascade_files <- file.path(wd, 'cascade.input.files.tsv')
out_array_hits <- file.path(wd, 'out_v3', 'cascade.array.hits.tsv')

gdrive_dir <- '1fGlnrj8Zu1Pox3fmxmFKdcbNVWiN9vhd'


traits <- fread(cascade_files)
dim(traits) %>% print()

#### read the variant annotation file for array

annot.arr <- read_annot_arr(annot.tbl)


#### read the GWAS summary statistics for the array

array_anno_unfiltered_df <- traits %>% 
read_array_sumstats_all(p_thr) %>%
annotate_array_df(annot.arr)


# quick check
array_anno_unfiltered_df %>% 
count(
    (as.numeric(P) < p_thr),
    is_autosome
)


array_anno_unfiltered_df %>% 
count(
    (as.numeric(maf) < maf_min_thr),    
    all_filters,
    is_outside_of_MHC,
    ld_indep
)


array_anno_df <- array_anno_unfiltered_df %>%
filter(
    (as.numeric(P) < p_thr),
    is_autosome,
    (as.numeric(maf) > maf_min_thr),
    all_filters == 0,
    is_outside_of_MHC,
    ld_indep    
) %>%
select(
    -is_autosome,
    -filter,
    -hwe,
    -gnomad_af,
    -missingness,
    -mcpi
)


# quick check
array_anno_df %>% count(all_filters, is_outside_of_MHC, ld_indep)


array_anno_df %>% 
select(Csq, is_rare, ID) %>%
unique() %>%
count(Csq, is_rare)


array_anno_df %>%
fwrite(out_array_hits, sep='\t')


paste('gdrive', 'upload', '-p', gdrive_dir, out_array_hits, sep=' ')


paste('zstd', '--rm', '-15', out_array_hits, sep=' ')
