==> cascade/4_cascade_imp.hits.ipynb <==
source('cascade_helper.R')


p_thr <- 5e-9


outd <- file.path(getwd(), 'out_v3')


traits <- fread('cascade.input.files.tsv') %>% drop_na(c('imp')) %>% filter(imp != '')


traits %>% dim()

imp_anno_df <- traits %>% 
read_imp_sumstats_all(p_thr) %>%
annotate_imp_df()


imp_anno_df %>% count(CHROM) %>% arrange(suppressWarnings(as.numeric(CHROM)), CHROM)

imp_anno_df %>% count(name)

imp_anno_df %>%
fwrite(file.path(outd, 'cascade.imp.hits.auto_and_X.tsv'), sep='\t')


imp_anno_df %>%
filter(CHROM != 'X') %>%
fwrite(file.path(outd, 'cascade.imp.hits.tsv'), sep='\t')


imp_anno_df %>% 
mutate(is_X = (CHROM == 'X')) %>%
select(is_rare, is_outside_of_MHC, is_X, ID) %>%
unique() %>%
count(is_X, is_outside_of_MHC, is_rare)

```
gdrive upload -p 1fGlnrj8Zu1Pox3fmxmFKdcbNVWiN9vhd out_v3/cascade.imp.hits.tsv
zstd -15 --rm out_v3/cascade.imp.hits.tsv
```
