==> cascade/6_number_check.ipynb <==
suppressWarnings(suppressPackageStartupMessages({
    library(tidyverse)
    library(data.table)
}))


# input file/dir
in_d <- '@@@@@@/users/ytanigaw/repos/rivas-lab/biomarkers/cascade/out_v3'


dfs <- list()
for(vtype in c('PTVs', 'PAVs', 'nc')){
    dfs[[vtype]] <- fread(file.path(in_d, sprintf('cascade.%s.tsv', vtype))) %>% rename('CHROM'='#CHROM')
}


df <- bind_rows(
    dfs[['PTVs']] %>% mutate(Csq='PTVs'),
    dfs[['PAVs']] %>% mutate(Csq='PAVs')
) %>%
filter(is_outside_of_MHC) %>%
select(CHROM, POS, REF, ALT, ID, trait, Gene_symbol, BETA, MAF, Csq, is_outside_of_MHC, P) %>%
bind_rows(
    dfs[['nc']] %>% filter(is_outside_of_MHC) %>%
    replace_na(list(Gene_symbol=''))%>%
    select(CHROM, POS, REF, ALT, ID, trait, BETA, MAF, is_outside_of_MHC, P) %>%
    mutate(Gene_symbol='', Csq='nc')
)


df %>% colnames()

### the number of associations

df %>% 
mutate(is_rare = if_else(MAF < 0.01, 'rare', 'non_rare')) %>%
count(is_outside_of_MHC, Csq, is_rare) %>%
spread(is_rare, n, fill=0) %>%
mutate(n = rare + non_rare)


df %>% 
filter(abs(BETA) > .1) %>%
mutate(
    is_rare = if_else(MAF < 0.01, 'rare', 'non_rare'),
    sign = if_else(BETA > 0, '+', '-')
) %>%
count(is_outside_of_MHC, Csq, sign, is_rare) %>%
spread(is_rare, n, fill=0) %>%
mutate(n = rare + non_rare)


### the number of variants

df %>% select(CHROM, POS, REF, ALT, ID) %>% unique %>% nrow


df %>% pull(ID) %>% n_distinct


df %>% filter(is.na(ID))


df %>% 
select(CHROM, POS, ID, MAF, Csq, is_outside_of_MHC) %>%
unique() %>% 
mutate(is_rare = if_else(MAF < 0.01, 'rare', 'non_rare')) %>%
count(is_outside_of_MHC, Csq, is_rare) %>%
spread(is_rare, n, fill=0) %>%
mutate(n = rare + non_rare)


5162 + 594 + 38

df %>% 
filter(abs(BETA) > .1) %>%
select(CHROM, POS, ID, MAF, Csq, is_outside_of_MHC) %>%
unique() %>% 
mutate(
    is_rare = if_else(MAF < 0.01, 'rare', 'non_rare'),
) %>%
count(is_outside_of_MHC, Csq, is_rare) %>%
spread(is_rare, n, fill=0) %>%
mutate(n = rare + non_rare)


df %>% count(trait)

df %>% select(CHROM, POS, REF, ALT, ID) %>% unique %>%
arrange(CHROM, POS, ID) %>%
rename('#CHROM' = 'CHROM') %>%
fwrite(file.path(in_d, 'list_of_loci.tsv'), sep='\t', na = "NA", quote=F)
